refactor(rp): split config.rs into per-domain config/ module
Mechanical move with no behavior change. Split the 852-line
services/rp/src/config.rs into a per-domain config/ subtree:

  config/
    mod.rs               Top-level Config struct, load_config(),
                         re-exports of all per-domain types so
                         existing crate::config::CameraConfig
                         imports keep working unchanged.
    site.rs              SiteConfig + impl validate + tests
    session.rs           SessionConfig + tests
    equipment.rs         EquipmentConfig wrapper
    camera.rs            CameraConfig + impl validate + tests
    focuser.rs           FocuserConfig + tests
    mount.rs             MountConfig + tests
    filter_wheel.rs      FilterWheelConfig
    cover_calibrator.rs  CoverCalibratorConfig + default fn
    imaging.rs           ImagingConfig + Default + default fns + tests
    plate_solver.rs      PlateSolverConfig + default fn + tests
    server.rs            ServerConfig + default_port/default_bind

Tests are colocated with their type, matching the imaging/ convention.
The shared MINIMAL_CONFIG_JSON fixture lives in a #[cfg(test)] gated
config::test_support submodule so each per-type test module pulls
it in via crate::config::test_support::MINIMAL_CONFIG_JSON.

Validation: cargo build passes, all 386 rp library tests still pass,
cargo fmt clean. No external callsites touched — re-exports preserve
the public surface.

First in a 3-step refactor sequence: config.rs (this), then
equipment.rs (per-device-type split per rp.md design doc), then
mcp.rs (per-tool-category split using rmcp 1.4 multi-router pattern).
