+++ services/rp/src/config/filter_wheel.rs
@@ -1,0 +1,16 @@
+use serde::Deserialize;
+
+#[derive(Debug, Clone, Deserialize)]
+pub struct FilterWheelConfig {
+    pub id: String,
+    #[serde(default)]
+    pub camera_id: String,
+    pub alpaca_url: String,
+    #[serde(default)]
+    pub device_number: u32,
+    #[serde(default)]
+    pub filters: Vec<String>,
+    /// Optional HTTP Basic Auth credentials for connecting to auth-enabled Alpaca services
+    #[serde(default)]
+    pub auth: Option<rp_auth::config::ClientAuthConfig>,
+}
