+++ services/rp/src/config/server.rs
@@ -1,0 +1,21 @@
+use serde::Deserialize;
+
+#[derive(Debug, Clone, Deserialize)]
+pub struct ServerConfig {
+    #[serde(default = "default_port")]
+    pub port: u16,
+    #[serde(default = "default_bind")]
+    pub bind_address: String,
+    #[serde(default)]
+    pub tls: Option<rp_tls::config::TlsConfig>,
+    #[serde(default)]
+    pub auth: Option<rp_auth::config::AuthConfig>,
+}
+
+fn default_port() -> u16 {
+    11115
+}
+
+fn default_bind() -> String {
+    "127.0.0.1".to_string()
+}
